refactor: Package 구조 변경, add product table
