+++ src/main/java/kitchenpos/order/domain/Order.java
@@ -14,6 +14,7 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
+import kitchenpos.table.domain.OrderTable;
 
 @Entity
 @Table(name = "orders")
