+++ src/main/java/kitchenpos/order/domain/OrderRepository.java
@@ -1,6 +1,7 @@
 package kitchenpos.order.domain;
 
 import java.util.List;
+import kitchenpos.table.domain.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
