+++ src/main/java/kitchenpos/product/dto/ProductRequest.java
@@ -1,7 +1,7 @@
-package kitchenpos.menu.dto;
+package kitchenpos.product.dto;
 
 import java.math.BigDecimal;
-import kitchenpos.menu.domain.Product;
+import kitchenpos.product.domain.Product;
 
 public class ProductRequest {
