+++ src/main/java/kitchenpos/table/domain/TableGroupCreateValidator.java
@@ -1,8 +1,7 @@
-package kitchenpos.order.domain;
+package kitchenpos.table.domain;
 
 import java.util.List;
 import java.util.Objects;
-import kitchenpos.order.domain.OrderTable;
 import org.springframework.util.CollectionUtils;
 
 public class TableGroupCreateValidator {
