+++ src/main/java/kitchenpos/table/domain/TableGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.table.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
