+++ src/main/java/kitchenpos/table/domain/TableGroupUnGroupValidator.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.table.domain;
 
 public class TableGroupUnGroupValidator {
