+++ src/main/java/kitchenpos/table/dto/TableGroupResponse.java
@@ -1,9 +1,9 @@
-package kitchenpos.order.dto;
+package kitchenpos.table.dto;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.order.domain.TableGroup;
+import kitchenpos.table.domain.TableGroup;
 
 public class TableGroupResponse {
