+++ src/test/java/kitchenpos/menu/domain/MenuProductTest.java
@@ -1,5 +1,7 @@
 package kitchenpos.menu.domain;
 
+import kitchenpos.product.domain.Product;
+
 public class MenuProductTest {
 
     public static MenuProduct 메뉴상품_생성(Long seq, Menu menu, Product product, long quantity) {
