+++ src/test/java/kitchenpos/order/acceptance/OrderAcceptanceTest.java
@@ -23,8 +23,8 @@ import kitchenpos.menu.dto.MenuGroupResponse;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.dto.OrderLineItemRequest;
 import kitchenpos.order.dto.OrderResponse;
-import kitchenpos.menu.dto.ProductResponse;
-import kitchenpos.order.dto.OrderTableResponse;
+import kitchenpos.product.dto.ProductResponse;
+import kitchenpos.table.dto.OrderTableResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
