+++ src/test/java/kitchenpos/product/dto/ProductResponseTest.java
@@ -4,8 +4,7 @@ import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.menu.dto.ProductResponse;
-import kitchenpos.menu.domain.Product;
+import kitchenpos.product.domain.Product;
 
 public class ProductResponseTest {
