+++ src/test/java/kitchenpos/table/domain/OrderTableTest.java
@@ -1,8 +1,5 @@
 package kitchenpos.table.domain;
 
-import kitchenpos.order.domain.OrderTable;
-import kitchenpos.order.domain.TableGroup;
-
 public class OrderTableTest {
 
     public static OrderTable 주문_테이블_생성(Long id, TableGroup tableGroup, int numberOfGuests, boolean empty) {
