+++ src/test/java/kitchenpos/table/dto/OrderTableRequestTest.java
@@ -1,7 +1,5 @@
 package kitchenpos.table.dto;
 
-import kitchenpos.order.dto.OrderTableRequest;
-
 public class OrderTableRequestTest {
 
     public static OrderTableRequest 주문_테이블_생성_요청_객체_생성(int numberOfGuests, boolean empty) {
