+++ src/test/java/kitchenpos/table/dto/OrderTableResponseTest.java
@@ -3,8 +3,7 @@ package kitchenpos.table.dto;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.order.dto.OrderTableResponse;
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 
 public class OrderTableResponseTest {
