+++ src/test/java/kitchenpos/tablegroup/dto/TableGroupRequestTest.java
@@ -3,8 +3,8 @@ package kitchenpos.tablegroup.dto;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.order.dto.TableGroupRequest;
-import kitchenpos.order.dto.TableGroupRequest.OrderTableDto;
+import kitchenpos.table.dto.TableGroupRequest;
+import kitchenpos.table.dto.TableGroupRequest.OrderTableDto;
 
 public class TableGroupRequestTest {
